Removed obsolete files
