--- exasol/ds/sandbox/runtime/ansible/ansible_runner_wrapper_docker_playbook.yml
@@ -1,44 +1,0 @@
-- name: Prepare environment
-  hosts: localhost
-  gather_facts: false
-  vars:
-    ansible_python_interpreter: python3
-  tasks:
-    - name: Add Docker Container to Ansible Inventory
-      add_host:
-        name: "{{docker_container}}"
-        groups: docker_container_group
-        ansible_connection: docker
-
-- name: Prepare Ansible Access
-  hosts: docker_container_group
-  gather_facts: true
-  vars:
-    my_ansible_user: "ubuntu"
-    jupyter_user: "jupyter"
-    jupyter_group: "jupyter"
-    jupyter_user_home: "/home/jupyter"
-    need_sudo: no
-  tasks:
-    - import_tasks: apt_update.yml
-    - name: Ansible Access
-      include_role:
-        name: ansible_access
-
-- name: Setup AI Lab Docker Container
-  hosts: docker_container_group
-  remote_user: "ubuntu"
-  gather_facts: true
-  vars:
-    ansible_python_interpreter: python3
-    user_name: jupyter
-    user_group: jupyter
-    docker_group: docker
-    user_home: /home/jupyter
-    initial_notebook_folder: "{{ user_home }}/notebook-defaults"
-    work_in_progress_notebooks: "{{ work_in_progress_notebooks }}"
-    need_sudo: yes
-    docker_integration_test: true
-  tasks:
-    - import_tasks: general_setup_tasks.yml
-    - import_tasks: cleanup_tasks.yml
