--- exasol/ds/sandbox/runtime/ansible/general_setup_tasks.yml
@@ -1,59 +1,0 @@
-- name: Set facts for Entry Point
-  vars:
-    jupyter_virtualenv: "{{user_home}}/jupyterenv"
-  ansible.builtin.set_fact:
-    dss_facts:
-      entrypoint: "{{user_home}}/entrypoint.py"
-      docker_user: "{{ ansible_user }}"
-      docker_group: "{{ docker_group }}"
-      jupyter:
-        virtualenv: "{{ jupyter_virtualenv }}"
-        server: "{{ jupyter_virtualenv }}/bin/jupyter-lab"
-        core: "{{ jupyter_virtualenv }}/bin/jupyter"
-        port: "49494"
-        user: "{{ user_name }}"
-        group: "{{ user_group }}"
-        home: "{{ user_home }}"
-        password: "{{ lookup('ansible.builtin.env', 'JUPYTER_LAB_PASSWORD', default='ailab') }}"
-        logfile: "{{ user_home }}/jupyter-server.log"
-      notebook_folder:
-        initial: "{{ initial_notebook_folder }}"
-        final: "{{user_home}}/notebooks"
-    cacheable: yes
-- name: Install rsync
-  include_role:
-    name: rsync
-- name: Create Group and User for Running Jupyter Server
-  include_role:
-    name: jupyter_user
-- name: Copy Entry Point Script
-  include_role:
-    name: entrypoint
-- name: Install Jupyter
-  include_role:
-    name: jupyter
-  vars:
-    jupyterlab_virtualenv: "{{ dss_facts.jupyter.virtualenv }}"
-    jupyterlab_command: "{{ dss_facts.jupyter.server }}"
-    jupyterlab_port: "{{ dss_facts.jupyter.port }}"
-    jupyterlab_password: "{{ dss_facts.jupyter.password }}"
-    jupyterlab_notebook_folder_initial: "{{ dss_facts.notebook_folder.initial }}"
-    jupyterlab_notebook_folder: "{{ dss_facts.notebook_folder.final }}"
-    work_in_progress_notebooks: "{{ work_in_progress_notebooks }}"
-- name: Change Owner of All Files and Dirs in Home Directory
-  ansible.builtin.file:
-    path: "{{ user_home }}"
-    owner: "{{ user_name }}"
-    group: "{{ user_group }}"
-    recurse: true
-  become: "{{need_sudo}}"
-- name: Clear pip Cache
-  ansible.builtin.file:
-    path: /root/.cache/pip
-    state: absent
-- name: Install Docker
-  include_role:
-    name: docker
-- name: Disable Core Dumps
-  include_role:
-    name: coredumps
