--- exasol/ds/sandbox/runtime/ansible/roles/jupyter/tasks/main.yml
@@ -1,48 +1,0 @@
----
-
-
-- name: Setup Jupyter
-  block:
-    - name: Install dependent apt packages
-      apt:
-        name: "{{apt_dependencies}}"
-        state: present
-        install_recommends: false
-      become: "{{need_sudo}}"
-
-    - name: Update pip version
-      include_tasks:
-        file: update-pip.yml
-
-    - name: Install JupyterLab and required packages
-      include_tasks:
-        file: pip-install.yml
-      vars:
-        requirements_file: "jupyter_requirements.txt"
-
-    - name: Setup Jupyterlab
-      ansible.builtin.import_tasks: jupyterlab.yml
-
-    - name: Install packages to be used inside Jupyter notebooks
-      include_tasks:
-        file: pip-install.yml
-      vars:
-        requirements_file: "notebook_requirements.txt"
-
-    - name: Copy tutorial notebook
-      ansible.builtin.import_tasks: tutorial.yml
-
-    - name: Change owner of all files and dirs in home directory
-      ansible.builtin.file:
-        path: "{{ user_home }}"
-        owner: "{{ user_name }}"
-        group: "{{ user_group }}"
-        recurse: true
-      become: "{{ need_sudo }}"
-
-    - name: Setup Systemd service
-      ansible.builtin.import_tasks: systemd.yml
-
-    - name: Setup motd
-      ansible.builtin.import_tasks: motd.yml
-
